q2/requirements.txt
# Core Framework
streamlit==1.29.0
fastapi==0.104.1
uvicorn==0.24.0

# LangChain ecosystem
langchain==0.1.0
langchain-community==0.0.10
langchain-groq==0.0.1
langgraph==0.0.20
langsmith==0.0.69

# Vector Database and Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# API clients
requests==2.31.0
alpha-vantage==2.3.1
newsapi-python==0.2.6

# Data processing
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0

# Rate limiting
slowapi==0.1.9

# Utilities
pydantic==2.5.0
python-multipart==0.0.6 